requirements.txt
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
python-dotenv==1.0.1
